refactor(codegen): move pairwise into lib
