+++ scripts/codegen/src/lib/pairwise.test.ts
