+++ scripts/codegen/src/lib/pairwise.ts
